Updates to riak client
