--- src/examples/com/basho/riakcs/client/CSCredentials.Riak.properties
@@ -1,7 +1,0 @@
-#Credentials for RiakCS Installation
-accessKey=your_cs_access_key
-secretKey=your_cs_secret_key
-
-#hostname:port for your RiakCS installation
-endPoint=localhost:8080
-useHttps=false
