+++ src/main/java/com/basho/riakcs/client/api/RiakCSException.java
