+++ src/main/java/com/basho/riakcs/client/api/RiakCSObjectManagement.java
