+++ src/main/java/com/basho/riakcs/client/api/RiakCSUserManagement.java
