+++ src/main/java/com/basho/riakcs/client/impl/CodecUtils.java
