+++ src/main/java/com/basho/riakcs/client/impl/CommunicationLayer.java
