+++ src/main/java/com/basho/riakcs/client/impl/XMLUtils.java
