+++ src/main/java/examples/com/basho/riakcs/client/BucketOperations.java
