+++ src/main/java/examples/com/basho/riakcs/client/CSCredentials.java
