+++ src/main/java/examples/com/basho/riakcs/client/Playground.java
