+++ src/main/resources/examples/com/basho/riakcs/client/CSCredentials.Riak.properties
@@ -1,0 +1,7 @@
+#Credentials for RiakCS Installation
+accessKey=L4UQLXHPIH44EQMAGIF3
+secretKey=UdgN8s-pzYBMN6y1askCWssuva0o5YQmtciWHg==
+
+#hostname:port for your RiakCS installation
+endPoint=10.73.71.186:8080
+useHttps=false
